move serde code to impl_serde
